Cell Phenotype Network Plot

This function generates a network plot to visualize cell phenotype interactions based on the provided interaction data. The plot highlights significant interactions with edges colored and weighted according to the p-values and odds ratios (OR).

Usage

Cell_phenotype_network_plot(
  cwas.test = NULL,
  best.model = NULL,
  interaction.best.model = NULL,
  cell.color = c("#cd3023", "#b91e45", "#f9bcbd", "gray", "#9BBBE1", "#90C4E9",
    "#8095CE"),
  cell.size = c(25, 20, 15, 14),
  edge.width = c(2, 3, 4),
  arrow.size = 0.75,
  arrow.width = 1.5,
  edge.color = c("darkred", "#b91e45", "#f9bcbd", "#9BBBE1", "#4198b9", "#2d4e76"),
  Interation.effect = T
)

Arguments

cwas.test

Data frame containing CWAS test data. If best.model is provided, this parameter is ignored.

best.model

List containing the best model information. It should have a component Chosen_model_info which is a data frame with columns celltype, Pvalue, and Beta. Default is NULL.

interaction.best.model

List containing the best interaction model information. It should have a component Chosen_model_info which is a data frame with columns celltype, Pvalue, and Beta. Default is NULL.

cell.color

Character vector specifying the colors for the cells. Default is c("#cd3023", "#b91e45", "#f9bcbd", "gray", "#9BBBE1", "#90C4E9", "#8095CE").

cell.size

Numeric vector specifying the sizes for the cells. Default is c(25, 20, 15, 14).

edge.width

Numeric vector specifying the widths of the edges for different significance levels. Default is c(2, 3, 4).

arrow.size

Numeric value specifying the size of the arrows. Default is 0.75.

arrow.width

Numeric value specifying the width of the arrows. Default is 1.5.

edge.color

Character vector specifying the colors for the edges based on p-values and OR. Default is c("darkred", "#b91e45", "#f9bcbd", "#9BBBE1", "#4198b9", "#2d4e76").

Interation.effect

Logical value indicating whether to use interaction effect or moderation effect. Default is TRUE.

Value

A plot of the cell phenotype network.